requirements.txt
# Netmiko for SSH connection to Cisco devices
netmiko>=4.3.0

# MCP SDK for building MCP server
mcp>=0.9.0

# Additional utilities
python-dotenv>=1.0.0
pydantic>=2.0.0